requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-tenant-schemas==1.12.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-environ==0.11.2